+++ src/Blockcore.Indexer.Cirrus/Controllers/CirrusQueryController.cs
@@ -38,6 +38,20 @@ namespace Blockcore.Indexer.Cirrus.Controllers
       }
 
       [HttpGet]
+      [Route("contract/list")]
+      public IActionResult GetGroupedContracts()
+      {
+         return OkPaging(cirrusMongoData.GroupedContracts());
+      }
+
+      [HttpGet]
+      [Route("contract/list/{contractType}")]
+      public IActionResult GetContracts([MinLength(2)][MaxLength(100)] string contractType, [Range(0, long.MaxValue)] int? offset = 0, [Range(1, 50)] int limit = 10)
+      {
+         return OkPaging(cirrusMongoData.ListContracts(contractType, offset, limit));
+      }
+
+      [HttpGet]
       [Route("{address}/assets")]
       public IActionResult GetAddressAssets([MinLength(30)][MaxLength(100)] string address, [Range(0, long.MaxValue)] int? offset = 0, [Range(1, 50)] int limit = 10)
       {
@@ -95,11 +109,43 @@ namespace Blockcore.Indexer.Cirrus.Controllers
       }
 
       [HttpGet]
-      [Route("contract/StandardToken/{address}")]
+      [Route("contract/standardtoken/{address}")]
       [SlowRequestsFilteerAttribute]
       public async Task<IActionResult> GetStandardTokenContractByAddress([MinLength(30)][MaxLength(100)] string address)
       {
          var contract = await standardTokenService.ComputeSmartContractForAddressAsync(address);
+
+         if (contract is null)
+         {
+            return NotFound();
+         }
+
+         return Ok(contract);
+      }
+
+      [HttpGet]
+      [Route("contract/standardtoken/{address}/{filterAddress}")]
+      [SlowRequestsFilteerAttribute]
+      public async Task<IActionResult> GetStandardTokenContractByAddress([MinLength(30)][MaxLength(100)] string address, [MinLength(30)][MaxLength(100)] string filterAddress)
+      {
+         var contract = await standardTokenService.ComputeSmartContractForAddressAsync(address);
+
+         if (contract is null)
+         {
+            return NotFound();
+         }
+
+         contract.TokenHolders = contract.TokenHolders.Where(t => t.Address == filterAddress).ToList();
+
+         return Ok(contract);
+      }
+
+      [HttpGet]
+      [Route("contract/nonfungibletoken/{address}")]
+      [SlowRequestsFilteerAttribute]
+      public async Task<IActionResult> GetNonFungibleTokenContractByAddress([MinLength(30)][MaxLength(100)] string address)
+      {
+         var contract = await nonFungibleTokenService.ComputeSmartContractForAddressAsync(address);
 
          if (contract is null)
          {