+++ src/Blockcore.Indexer.Cirrus/Storage/ICirrusStorage.cs
@@ -10,6 +10,8 @@ public interface ICirrusStorage
    QueryResult<QueryContractCall> ContractCall(string address, string filterAddress, int? offset, int limit);
    QueryContractTransaction ContractTransaction(string transacitonId);
    QueryContractCode ContractCode(string address);
+   QueryResult<QueryContractGroup> GroupedContracts();
+   QueryResult<QueryContractList> ListContracts(string contractType, int? offset, int limit);
 
    Task<QueryResult<QueryAddressAsset>> GetAssetsForAddressAsync(string address, int? offset, int limit);
 }